backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==10.1.0
pillow-heif==0.13.0
opencv-python==4.8.1.78
numpy==1.26.4
easyocr==1.7.0
torch==2.2.0
torchvision==0.17.0
spacy==3.7.2
nltk==3.8.1
fuzzywuzzy==0.18.0
Levenshtein==0.25.0